requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
requests==2.31.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
Werkzeug==2.3.7
beautifulsoup4==4.12.2
APScheduler==3.10.1
Flask-Mail==0.9.1
